deploy/mt5-api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.9
pydantic==1.10.11
python-dotenv==1.0.0